Move splice tests to it's own file
